[Feat #27] 동네생활 커뮤니티 댓글 삭제 api연동

+++ app/src/main/java/com/daepiro/numberoneproject/data/model/CommunityTownReplyResponseItem.kt
@@ -5,7 +5,7 @@ data class CommunityTownReplyResponseItem(
     val authorNickName: String="",
     val authorProfileImageUrl: String="",
     val childComments: List<Any>?,
-    val commentId: Long=0,
+    val commentId: Int=0,
     val content: String = "",
     val createdAt: String = "",
     val likeCount: Int = 0,
